+++ packages/Angular/Generic/forms/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @memberjunction/ng-forms
 
+## 5.48.0
+
+### Patch Changes
+
+- @memberjunction/ai-core-plus@5.48.0
+- @memberjunction/ng-markdown@5.48.0
+- @memberjunction/ng-ui-components@5.48.0
+
 ## 5.47.0
 
 ### Patch Changes