+++ packages/Config/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @memberjunction/config
 
+## 5.48.0
+
 ## 5.47.0
 
 ## 5.46.0